+++ docs/changelog.rst
@@ -3,9 +3,17 @@ Changelog
 
 NEXT
 ----
+Bug fixes
+^^^^^^^^^
+
 * Fix error when Django happens to be imported before pytest-django runs.
   Thanks to Will Harris for `the bug report
   <https://github.com/pytest-dev/pytest-django/issues/289>`_.
+
+Features
+^^^^^^^^
+* Added a new option `--migrations` to negate a default usage of
+  `--nomigrations`.
 
 2.9.1
 -----